Fixes #15345 - Move RateControl to jetty-io.
Introduced oej.io.RateControl and retrofitted oej.http2.RateControl for backward compatibility.
