+++ jetty-core/jetty-http2/jetty-http2-server/src/main/config/etc/jetty-http2c.xml
@@ -11,7 +11,7 @@
         <Set name="initialSessionRecvWindow" property="jetty.http2c.initialSessionRecvWindow"/>
         <Set name="maxSettingsKeys" property="jetty.http2c.maxSettingsKeys"/>
         <Set name="rateControlFactory">
-          <New class="org.eclipse.jetty.http2.WindowRateControl$Factory">
+          <New class="org.eclipse.jetty.io.WindowRateControl$Factory">
             <Arg type="int"><Property name="jetty.http2c.rateControl.maxEventsPerSecond" default="128"/></Arg>
           </New>
         </Set>
