+++ jetty-ee10/jetty-ee10-osgi/test-jetty-ee10-osgi/src/test/config/etc/jetty-http2.xml
@@ -11,7 +11,7 @@
         <Set name="initialSessionRecvWindow" property="jetty.http2.initialSessionRecvWindow"/>
         <Set name="maxSettingsKeys"><Property name="jetty.http2.maxSettingsKeys" default="64"/></Set>
         <Set name="rateControlFactory">
-          <New class="org.eclipse.jetty.http2.WindowRateControl$Factory">
+          <New class="org.eclipse.jetty.io.WindowRateControl$Factory">
             <Arg type="int"><Property name="jetty.http2.rateControl.maxEventsPerSecond" default="50"/></Arg>
           </New>
         </Set>
