+++ jetty-ee11/jetty-ee11-tests/jetty-ee11-test-http2-webapp/src/test/java/org/eclipse/jetty/test/webapp/StressHTTP2AbortTest.java
@@ -25,12 +25,12 @@ import java.util.concurrent.TimeoutException;
 import org.eclipse.jetty.client.HttpClient;
 import org.eclipse.jetty.client.Request;
 import org.eclipse.jetty.http.HttpMethod;
-import org.eclipse.jetty.http2.RateControl;
-import org.eclipse.jetty.http2.WindowRateControl;
 import org.eclipse.jetty.http2.client.HTTP2Client;
 import org.eclipse.jetty.http2.client.transport.HttpClientTransportOverHTTP2;
 import org.eclipse.jetty.http2.server.HTTP2CServerConnectionFactory;
 import org.eclipse.jetty.io.EndPoint;
+import org.eclipse.jetty.io.RateControl;
+import org.eclipse.jetty.io.WindowRateControl;
 import org.eclipse.jetty.server.Handler;
 import org.eclipse.jetty.server.HttpConfiguration;
 import org.eclipse.jetty.server.Response;
